fix(fdf): restructure color parsing
