+++ extra-make/sources.mk
@@ -6,7 +6,7 @@
 #    By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/16 16:05:45 by elagouch          #+#    #+#              #
-#    Updated: 2025/03/05 13:23:07 by elagouch         ###   ########.fr        #
+#    Updated: 2025/03/06 10:19:28 by elagouch         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -164,6 +164,7 @@ _SRC_FDF_COMMON				:=	\
 	color_get					\
 	color_get_by_scheme			\
 	color_get_line				\
+	color_parsing				\
 	color_utils					\
 	debug_toggle				\
 	error						\
@@ -188,28 +189,28 @@ _SRC_FDF_COMMON				:=	\
 	mlx_hooks					\
 	mlx_hooks_key				\
 	mlx_pixel_put				\
+	mlx_render_line				\
 	mlx_render_lod				\
 	mlx_render_lod_debug		\
 	mlx_render_lod_debug_utils	\
-	mlx_render_line				\
+	mlx_render_next_frame		\
 	mlx_render_section			\
 	mlx_render_section_debug	\
 	mlx_render_section_utils	\
 	mlx_render_viewport			\
-	mlx_render_next_frame		\
+	point3d_rotate				\
 	point_add					\
+	point_cabinet_project		\
+	point_conic_project			\
 	point_distance				\
 	point_dot					\
 	point_iso_project			\
-	point_cabinet_project		\
 	point_lerp					\
 	point_magnitude				\
 	point_normalize				\
 	point_rotate				\
 	point_scale					\
-	point_sub					\
-	point_conic_project			\
-	point3d_rotate
+	point_sub
 _SRC_FDF_COMMON				:=	\
 	$(addprefix src/, $(_SRC_FDF_COMMON))
 _SRC_FDF_MANDATORY			:=	\
