+++ milestone-2-fdf/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "files.associations": {
+    "*.astro": "astro",
+    "fdf.h": "c"
+  }
+}
