+++ milestone-2-fdf/src/map_parse_utils.c
@@ -6,7 +6,7 @@
 /*   By: elagouch <elagouch@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/24 17:22:42 by elagouch          #+#    #+#             */
-/*   Updated: 2025/03/05 13:44:56 by elagouch         ###   ########.fr       */
+/*   Updated: 2025/03/06 10:20:35 by elagouch         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,39 +47,6 @@ int	count_columns_in_line(const char *line)
 }
 
 /**
- * @brief Convert a hex character to its integer value
- * @param c Hexadecimal character (0-9, a-f, A-F)
- * @return int Corresponding integer value (0-15)
- */
-int	hex_value(char c)
-{
-	if (c >= '0' && c <= '9')
-		return (c - '0');
-	else if (c >= 'a' && c <= 'f')
-		return (c - 'a' + 10);
-	else if (c >= 'A' && c <= 'F')
-		return (c - 'A' + 10);
-	return (0);
-}
-
-/**
- * Checks if a character is a hexadecimal digit (0-9, a-f, A-F)
- * @param c The character to check
- * @return 1 if the character is a hexadecimal digit, 0 otherwise
- */
-int	ft_ishexdigit(int c)
-{
-	int	is_digit;
-	int	is_lower_hex;
-	int	is_upper_hex;
-
-	is_digit = (c >= '0' && c <= '9');
-	is_lower_hex = (c >= 'a' && c <= 'f');
-	is_upper_hex = (c >= 'A' && c <= 'F');
-	return (is_digit || is_lower_hex || is_upper_hex);
-}
-
-/**
  * @brief Process a token to extract elevation and optional color
  *
  * @param ctx Application context
@@ -107,32 +74,4 @@ void	parse_token(t_app *ctx, char *token, int row, int *col)
 		ctx->map.matrix[row][*col].color = color;
 		(*col)++;
 	}
-}
-
-/**
- * @brief Parse color value from hex string
- *
- * @param ctx Application context
- * @param str Color string to parse
- * @param color Pointer to store parsed color
- * @return t_bool Success or failure
- */
-t_bool	parse_color(t_app *ctx, char *str, unsigned int *color)
-{
-	int	i;
-
-	(void)ctx;
-	if (str[0] != '0' || (str[1] != 'x' && str[1] != 'X'))
-		return (false);
-	str += 2;
-	*color = 0;
-	i = 0;
-	while (str[i] && ft_ishexdigit(str[i]))
-	{
-		*color = *color * 16 + (unsigned int)hex_value(str[i]);
-		i++;
-	}
-	if (str[i] && !ft_isspace(str[i]) && str[i] != '\0')
-		return (false);
-	return (true);
 }
